Don't create RethinkDB instance if it is already configured in this VM.
For example, from already having run ./rethinkdb.sh

+++ Vagrantfile
@@ -24,9 +24,12 @@ Vagrant::Config.run do |config|
         apt-get update --assume-yes;
         apt-get install --assume-yes rethinkdb;
 
-        sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
-                                echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
-        rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+        if [ ! -f /etc/rethinkdb/instances.d/default.conf ];
+        then
+          sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
+          echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
+          rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+        fi
 
         service rethinkdb start;
       EOF
@@ -60,10 +63,13 @@ Vagrant::Config.run do |config|
           apt-get update --assume-yes;
           apt-get install --assume-yes rethinkdb;
 
-          sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
-          echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
-          echo "join=10.10.10.10:29015" >> /etc/rethinkdb/instances.d/default.conf
-          rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+          if [ ! -f /etc/rethinkdb/instances.d/default.conf ];
+          then
+            sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
+            echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
+            echo "join=10.10.10.10:29015" >> /etc/rethinkdb/instances.d/default.conf
+            rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+          fi
 
           service rethinkdb start;
         EOF
